Increase rotator precision test to 1e-6
During testing, it appeared that the rotator was showing accuracy
to 1e-5 for the avx kernel.  (the generic kernel matched the CPU
calculations well, so I don't think there's any issue there).
This shorter precision was resulting in cycling noise of
about 10dB when tested in GNU Radio with a flowgraph consisting of
signal source->throttle->rotator->frequency sink.  While the noise
starts to appear at about -100 dB it is very prominent around
-120 dB.  If the rotator is used to frequency shift very weak
satellite or radio astronomy signals, this low-level noise
could potentially manifest in the integrated data.  This PR is
suggesting considering rotator accuracy to 1e-6 to trivialize
the resulting noise.